+++ examples/pixels/src/components/avatar/index.ts
@@ -1,0 +1,2 @@
+export * from "./avatar-panel";
+export * from "./contributors";