+++ examples/pixels/src/components/layout/index.ts
@@ -1,0 +1,2 @@
+export { Header } from "./header";
+export { Layout } from "./layout";