+++ examples/pixels/src/utility/constants.ts
@@ -1,0 +1,21 @@
+export const PIXEL_SIZE = 10;
+export const DEFAULT_SCALE = 1;
+export const DEFAULT_CANVAS_SIZE = 30;
+export const colors = [
+    "black",
+    "white",
+    "dimgray",
+    "sienna",
+    "tan",
+    "navy",
+    "cornflowerblue",
+    "teal",
+    "darkolivegreen",
+    "mediumseagreen",
+    "purple",
+    "plum",
+    "gold",
+    "darkorange",
+    "orangered",
+    "firebrick",
+] as const;