+++ examples/pixels/src/utility/get-user-query-ids.ts
@@ -1,0 +1,10 @@
+import { Pixel } from "types/pixel";
+
+export const getUserQueryIds = (pixels: Pixel[] | undefined): string[] => {
+    const ids: string[] | undefined = pixels?.map(
+        (pixel: any) => pixel.user_id,
+    );
+    const idSet = new Set(ids);
+
+    return [...idSet];
+};