+++ packages/inferencer/src/field-transformers/index.ts
@@ -1,3 +1,4 @@
+import { basicToRelation } from "./basic-to-relation";
 import { imageByKey } from "./image-by-key";
 import { relationByResource } from "./relation-by-resource";
 import { relationToFieldable } from "./relation-to-fieldable";
@@ -6,4 +7,5 @@ export const defaultTransformers = [
     imageByKey,
     relationByResource,
     relationToFieldable,
+    basicToRelation,
 ];