+++ packages/inferencer/src/utilities/remove-relation-suffix/index.ts
@@ -1,7 +1,9 @@
+import { relationRegexp } from "@/field-inferencers/relation";
+
 /**
  * Removes the relation suffix from a string.
  * @example removeRelationSuffix("category_id") === "category"
  */
 export const removeRelationSuffix = (key: string) => {
-    return key.replace(/(_id|_ids|Id|Ids|ID|IDs)(\[\])?$/, "");
+    return key.replace(relationRegexp, "");
 };